Feature/42 adicionar jwt

+++ config/routes.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 use Minascafe\Category\Infrastructure\Http\Controller\CategoryController;
 use Minascafe\Product\Infrastructure\Http\Controller\ProductController;
+use Minascafe\User\Infrastructure\Http\Controller\SessionController;
 use Minascafe\User\Infrastructure\Http\Controller\UserController;
+use Minascafe\User\Infrastructure\Http\Middleware\AuthorizationMiddleware;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Slim\App;
@@ -17,20 +19,24 @@ return function (App $app) {
         return $response;
     });
 
+    $app->group('/autenticacao', function (RouteCollectorProxy $group) {
+        $group->post('', [SessionController::class, 'create']);
+    });
+
     $app->group('/categorias', function (RouteCollectorProxy $group) {
         $group->get('', [CategoryController::class, 'index']);
-        $group->post('', [CategoryController::class, 'create']);
-        $group->get('/{id}', [CategoryController::class, 'show']);
-        $group->put('/{id}', [CategoryController::class, 'update']);
-        $group->delete('/{id}', [CategoryController::class, 'destroy']);
+        $group->post('', [CategoryController::class, 'create'])->add(AuthorizationMiddleware::class);
+        $group->get('/{id}', [CategoryController::class, 'show'])->add(AuthorizationMiddleware::class);
+        $group->put('/{id}', [CategoryController::class, 'update'])->add(AuthorizationMiddleware::class);
+        $group->delete('/{id}', [CategoryController::class, 'destroy'])->add(AuthorizationMiddleware::class);
     });
 
     $app->group('/produtos', function (RouteCollectorProxy $group) {
         $group->get('', [ProductController::class, 'index']);
-        $group->post('', [ProductController::class, 'create']);
-        $group->get('/{id}', [ProductController::class, 'show']);
-        $group->put('/{id}', [ProductController::class, 'update']);
-        $group->delete('/{id}', [ProductController::class, 'destroy']);
+        $group->post('', [ProductController::class, 'create'])->add(AuthorizationMiddleware::class);
+        $group->get('/{id}', [ProductController::class, 'show'])->add(AuthorizationMiddleware::class);
+        $group->put('/{id}', [ProductController::class, 'update'])->add(AuthorizationMiddleware::class);
+        $group->delete('/{id}', [ProductController::class, 'destroy'])->add(AuthorizationMiddleware::class);
     });
 
     $app->group('/usuarios', function (RouteCollectorProxy $group) {
@@ -39,7 +45,7 @@ return function (App $app) {
         $group->get('/{id}', [UserController::class, 'show']);
         $group->put('/{id}', [UserController::class, 'update']);
         $group->delete('/{id}', [UserController::class, 'destroy']);
-    });
+    })->add(AuthorizationMiddleware::class);
 
     /*
     * Catch-all route to serve a 404 Not Found page if none of the routes match
